modularize realDiv
