+++ tfjs-core/src/ops/div.ts
@@ -16,7 +16,7 @@
  */
 
 import {ENGINE, ForwardFunc} from '../engine';
-import {Div, DivInputs} from '../kernel_names';
+import {RealDiv, RealDivInputs} from '../kernel_names';
 import {Tensor} from '../tensor';
 import {NamedTensorMap} from '../tensor_types';
 import {makeTypesMatch} from '../tensor_util';
@@ -65,12 +65,12 @@ function div_<T extends Tensor>(a: Tensor|TensorLike, b: Tensor|TensorLike): T {
     return res;
   };
 
-  const inputs: DivInputs = {a: $a, b: $b};
+  const inputs: RealDivInputs = {a: $a, b: $b};
   const attrs = {};
 
   return ENGINE.runKernelFunc(
-             forward, inputs as {} as NamedTensorMap, null /* gradient */, Div,
-             attrs) as T;
+             forward, inputs as {} as NamedTensorMap, null /* gradient */,
+             RealDiv, attrs) as T;
 }
 
 export const div = op({div_});
