+++ tfjs-core/src/register_all_gradients.ts
@@ -43,7 +43,6 @@ import {coshGradConfig} from './gradients/Cosh_grad';
 import {cumsumGradConfig} from './gradients/Cumsum_grad';
 import {depthwiseConv2dNativeGradConfig} from './gradients/DepthwiseConv2dNative_grad';
 import {dilation2dGradConfig} from './gradients/Dilation2D_grad';
-import {divGradConfig} from './gradients/Div_grad';
 import {eluGradConfig} from './gradients/Elu_grad';
 import {erfGradConfig} from './gradients/Erf_grad';
 import {expGradConfig} from './gradients/Exp_grad';
@@ -76,6 +75,7 @@ import {onesLikeGradConfig} from './gradients/OnesLike_grad';
 import {padV2GradConfig} from './gradients/PadV2_grad';
 import {powGradConfig} from './gradients/Pow_grad';
 import {preluGradConfig} from './gradients/Prelu_grad';
+import {divGradConfig} from './gradients/RealDiv_grad';
 import {reciprocalGradConfig} from './gradients/Reciprocal_grad';
 import {relu6GradConfig} from './gradients/Relu6_grad';
 import {reluGradConfig} from './gradients/Relu_grad';
